+++ src/main/java/dk/kb/util/xml/XMLUtil.java
@@ -53,7 +53,6 @@ import java.util.Objects;
  */
 public class XMLUtil {
 
-    static final Logger logger = LoggerFactory.getLogger(XMLUtil.class);
 
     private static final ThreadLocal<ReplaceReader> localEncoder =
             new ThreadLocal<ReplaceReader>() {
@@ -114,175 +113,6 @@ public class XMLUtil {
             default:
                 return "UNKNOWN_EVENT_TYPE " + "," + eventType;
         }
-    }
-
-
-    /**
-     * Creates an empty Document object with the specified XML version and whether it is standalone.
-     * @param xmlVersion The used XML version.
-     * @param standalone Whether it is standalone in the XML sense of the word
-     *          (see for example <a href="https://stackoverflow.com/questions/5578645/what-does-the-standalone-directive-mean-in-xml">"https://stackoverflow.com/questions/5578645/what-does-the-standalone-directive-mean-in-xml"</a>)
-     * @return A new empty Document object.
-     */
-    public static Document createDocument(String xmlVersion, boolean standalone) {
-        try {
-            DocumentBuilderFactory documentBuilderFactory = DocumentBuilderFactory.newInstance();
-            DocumentBuilder documentBuilder = documentBuilderFactory.newDocumentBuilder();
-            Document newDocument = documentBuilder.newDocument();
-            newDocument.setXmlVersion(xmlVersion);
-            newDocument.setXmlStandalone(standalone);
-            return newDocument;
-        } catch (ParserConfigurationException e) {
-            throw new XMLException("An error occurred while creating a DocumentBuilder for a new Document.", e);
-        }
-    }
-
-    /**
-     * This is the simpler version of {@link #createDocument(String, boolean)}
-     * where the values are set with default: xmlVersion="1.0" and standalone=true.
-     * @return A new empty Document object.
-     */
-    public static Document createDocument() {
-        return createDocument("1.0", true);
-    }
-
-    /**
-     * Takes an XML fragment represented as a String object and parses it to a Document object.
-     * @param xml The XML fragment represented as a string.
-     * @return A document object, parsed from the given XML string.
-     */
-    public static Document parseDocument(String xml) {
-        try {
-            DocumentBuilderFactory documentBuilderFactory = DocumentBuilderFactory.newInstance();
-            DocumentBuilder documentBuilder = documentBuilderFactory.newDocumentBuilder();
-            return documentBuilder.parse(new ByteArrayInputStream(xml.getBytes(StandardCharsets.UTF_8)));
-        } catch (ParserConfigurationException | SAXException | IOException e) {
-            throw new XMLException("An error occurred while parsing XML from string to a Document." + xml, e);
-        }
-    }
-
-    /**
-     * It is so to speak the reverse of {@link #parseDocument(String)}.
-     * @param document The document to be transformed into a string.
-     * @param omitXmlDeclaration Whether the returned string will omit/include the XML declaration header (<?xml version="1.0" ...?>)
-     *                           true->no XML declaration is added,
-     *                           false->string including XML declaration header.
-     * @return A string representation of the given Document.
-     */
-    public static String getStringFromDocument(Document document, boolean omitXmlDeclaration) {
-        try {
-            return DOM.domToString(document, !omitXmlDeclaration, true);
-        } catch (TransformerException e) {
-            throw new XMLException("An error occurred while trying to transform a document to a string.", e);
-        }
-
-    }
-
-    /**
-     * This is the simpler version of {@link #getStringFromDocument(Document, boolean)}
-     * where the boolean value for omitting XML declaration is set to 'true' (no XML declaration header).
-     * @param document The document to be transformed into a string.
-     * @return A string representation of the given Document.
-     */
-    public static String getStringFromDocument(Document document) {
-        return getStringFromDocument(document, true);
-    }
-
-    /**
-     * Takes a document and adds a root element to it with the specified tag name and attributes.
-     * @param document The document to which the specified root element is added.
-     * @param rootElementTagName The name of the new root element.
-     * @param xmlns The default namespace the document in which the fragment is being created.
-     * @param extraAttributes A map containing (key, value) pairs of extra attributes. This may be an empty map.
-     * @return The newly created root element with the following structure:
-     * <rootElementTagName xmlns="..." key1="value1" key2="value2">
-     *     ...
-     * </rootElementTagName>
-     */
-    public static Element addRootElementToDocument(Document document, String rootElementTagName,
-                                                   String xmlns, Map<String, String> extraAttributes) {
-        Element rootElement = document.createElement(rootElementTagName);
-        Objects.requireNonNull(xmlns, "Method requires 'xmlns' to be non-null, but it is.");
-        rootElement.setAttribute("xmlns", xmlns);
-        for (Map.Entry<String, String> entry : extraAttributes.entrySet()) {
-            rootElement.setAttribute(entry.getKey(), entry.getValue());
-        }
-        document.appendChild(rootElement);
-        return rootElement;
-    }
-
-    /**
-     * This is the simpler version of {@link #addRootElementToDocument(Document, String, String, Map)}
-     * where the map of extra attributes is empty (meaning no extra attributes to set).
-     * @param document The document to which the specified root element is added.
-     * @param rootElementTagName The name of the new root element.
-     * @param xmlns The default namespace the document in which the fragment is being created.
-     * @return The newly created root element with the following structure:
-     * <rootElementTagName xmlns="...">
-     *     ...
-     * </rootElementTagName>
-     */
-    public static Element addRootElementToDocument(Document document, String rootElementTagName, String xmlns) {
-        return addRootElementToDocument(document, rootElementTagName, xmlns, Map.of());
-    }
-
-    /**
-     * This is a method to add a child element with no content to an existing element.
-     * Similar to method {@link #addElementWithTextContent(Element, String, String)} which additionally also
-     * gives the new element some text content.
-     * @param parentElement The parent element of the newly created element.
-     * @param elementTagName The name of the newly created element: <elementTagName/>.
-     * @return The element object of the newly created element which has been added (appended) to the given parentElement.
-     */
-    public static Element addElement(Element parentElement, String elementTagName) {
-        Element newElement = parentElement.getOwnerDocument().createElement(elementTagName);
-        parentElement.appendChild(newElement);
-        return newElement;
-    }
-
-    /**
-     * This is a method to add a child element with som text content to an existing element.
-     * Similar to method {@link #addElement(Element, String)} except it does not give the new element
-     * any text content.
-     * @param parentElement The parent element of the newly created element.
-     * @param elementTagName The name of the newly created element: <elementTagName>...</elementTagName>.
-     * @param textContent The content of the newly created element: <elementTagName>textContent</elementTagName>
-     * @return The element object of the newly created element which has been added (appended) to the given parentElement.
-     */
-    public static Element addElementWithTextContent(Element parentElement, String elementTagName, String textContent) {
-        Element newElement = addElement(parentElement, elementTagName);
-        newElement.setTextContent(textContent);
-        return newElement;
-    }
-
-    /**
-     * Removes the element specified by 'elementTag' from the root element of the given XML.
-     * @param xmlString the XML, represented as a string, for which the specified element is to be removed from.
-     * @param elementTag the tag name of the element to be removed. Must be a child of the root element.
-     * @return If the specified element is found, the new XML (represented as a string) with the specified element removed.
-     * Otherwise, the initial xmlString given as argument is returned.
-     */
-    public static String removeElementFromXml(String xmlString, String elementTag) {
-        Document document = parseDocument(xmlString);
-        Element rootElement = document.getDocumentElement();
-        NodeList childElementNodeList = rootElement.getElementsByTagName(elementTag);
-        if (childElementNodeList.getLength() > 0) {
-            Node childElementNode = childElementNodeList.item(0);
-            rootElement.removeChild(childElementNode);
-            return getStringFromDocument(document);
-        } else {
-            logger.warn("No child element was found for tag {}. Returning original xml string.", elementTag);
-            return xmlString;
-        }
-    }
-
-    /**
-     * Takes a boolean value and returns a string of 'yes' or 'no' depending on the input.
-     * @param bool The value to convert/translate.
-     * @return A string representing the parameter input value: true->'yes' and false->'no'.
-     */
-    private static String booleanToYesNo(boolean bool) {
-        return bool ? "yes" : "no";
     }
 
 }
